feat: use toast context
